requirements.txt
PyYAML
colorama
pyfiglet
PyInquirer
python-binance
telethon